{{cookiecutter.repo_name}}/requirements/base.txt
django==1.6.2
django-braces==1.3.1
South==0.8.4
django-model-utils==2.0
Pillow==2.3.0
psycopg2==2.5.2
unicode-slugify==0.1.1
django-autoslug==1.7.2
